Refactor to more idiomatic layout
