+++ driver/config.go
@@ -109,40 +109,3 @@ var (
 		})),
 	})
 )
-
-// NetworkConfig describes networking configuration for a task
-type NetworkConfig struct {
-	// Mode selects networking mode: "host", "bridged", "softnet", or "shared" (default NAT)
-	Mode string `codec:"mode"`
-	// BridgedInterface is used when Mode == "bridged" to select the interface
-	BridgedInterface string `codec:"bridged_interface"`
-	// SoftnetAllow CIDRs when using Softnet; implies Softnet if Mode unspecified
-	SoftnetAllow []string `codec:"softnet_allow"`
-	// SoftnetExpose EXTERNAL:INTERNAL TCP port forward specs when using Softnet; implies Softnet
-	SoftnetExpose []string `codec:"softnet_expose"`
-}
-
-// Options to specify how the root disk of the VM should be
-// prepared by the tart tool.
-type RootDiskOptions struct {
-	ReadOnly    bool    `codec:"readonly"`
-	SyncMode    *string `codec:"sync_mode"`
-	CachingMode *string `codec:"caching_mode"`
-}
-
-// DirectoryMount represents a single directory block item from the config
-// with an optional name (purely descriptive), required host path, and
-// optional options.
-type DirectoryMount struct {
-	Name    string            `codec:"name"`
-	Path    string            `codec:"path"`
-	Options *DirectoryOptions `codec:"options"`
-}
-
-// DirectoryOptions controls how a directory mount is handled by tart.
-// - readonly: when true, append ":ro" to the mount spec
-// - tag: when set, append "@tag" to the mount spec
-type DirectoryOptions struct {
-	ReadOnly bool   `codec:"readonly"`
-	Tag      string `codec:"tag"`
-}
