--- driver/waiting.go
@@ -1,31 +1,0 @@
-package driver
-
-import (
-	"context"
-	"fmt"
-
-	"github.com/hashicorp/nomad/plugins/drivers"
-)
-
-func (d *Driver) handleWait(ctx context.Context, handle *taskHandle, ch chan *drivers.ExitResult) {
-	defer close(ch)
-
-	var result *drivers.ExitResult
-	ps, err := handle.exec.Wait(ctx)
-	if err != nil {
-		result = &drivers.ExitResult{
-			Err: fmt.Errorf("executor: error waiting on process: %v", err),
-		}
-	} else {
-		result = &drivers.ExitResult{
-			ExitCode: ps.ExitCode,
-			Signal:   ps.Signal,
-		}
-	}
-
-	select {
-	case <-ctx.Done():
-	case <-d.ctx.Done():
-	case ch <- result:
-	}
-}
